Update README.md
Face detection is a computer vision technology that helps to locate/visualize human faces in digital images. This technique is a specific use case of object detection technology that deals with detecting instances of semantic objects of a certain class (such as humans, buildings or cars) in digital images and videos. With the advent of technology, face detection has gained a lot of importance especially in fields like photography, security, and marketing, hands-on knowledge. When coming to speech production, it is a useful technology that converts any text into a speech signal and it has a lot of importance in car navigation ,email reading etc. Speech synthesis can be described as artificial production of human speech. In this project we provide voice output or speech production additionally when the face is recognized. After recognizing the faces the text gets converted into speech so that we get voice as output. Generally speech is one of the oldest and most natural means of information exchange between human. Now, the aim of this project is to exchange the information between human and machine and has to detect and recognize faces in a scene and to provide the user with the person's name as well as additional speech in the form of voice. We use machine learning algorithm to validate unspecified images by comparing them to previously saved pictures in the training set and to provide information about when the individual recognized by the system.